Rename validate_user.dto.ts to validate-user.dto.ts
